+++ src/types/interfaces.ts
@@ -1020,6 +1020,7 @@ export interface SdkEvents {
   handledPendingSession: ISessionIdAndConversationId;
   cancelPendingSession: ISessionIdAndConversationId;
   conversationUpdate: ISdkConversationUpdateEvent;
+  conversationUpdateRaw: SubscriptionEvent;
   station: (event: { action: 'Associated' | 'Disassociated', station: IStation | null }) => void;
   concurrentSoftphoneSessionsEnabled: boolean; // lineAppearence > 1
   resolutionUpdated: IResolutionChange
@@ -1287,8 +1288,6 @@ export interface IncomingStreamStatus {
   sourcePinId: string;
   /** Track ID from the SDP of the viewed member's client */
   sourceTrackId?: string;
-  //TODO: Remove this when the camelcase version is deployed for media service
-  appid?: IDataChannelAppId;
   appId?: IDataChannelAppId;
   contentType?: 'camera' | 'screenshare' | 'playback';
 }